fix(plugins): rename ldap/organization.ts to ldap/organizations.ts
The plugin loader resolves `core/<path>` directly to a JS file, so
loading this plugin with the documented name `core/ldap/organizations`
fails with ERR_MODULE_NOT_FOUND because the source file was named
`organization.ts` (singular). The class (`LdapOrganizations`),
internal plugin name (`ldapOrganizations`), browser API URLs
(`/api/v1/ldap/organizations/...`), the docs file
(`docs/usage/plugins/ldap/organizations.md`) and every reference in the
docs already use the plural form, so the singular filename was the
outlier. Renaming it to `organizations.ts` makes `core/ldap/organizations`
load as documented and aligns the source with the rest of the codebase.

The companion test file is renamed for symmetry, and the six test
imports that pointed at `'src/plugins/ldap/organization'` are updated.
No behavioural change; the full test suite still passes (703 passing,
7 pending).
